debug

+++ src/add-extra-sam-flags.cpp
@@ -264,7 +264,7 @@ main(int argc, char* argv[])
       if (local_m_vector.size() == 0)
         break;
 
-      for (i = 0; i < local_m_vector.size(); ++i) {
+      for (i = 0; i < (int)local_m_vector.size(); ++i) {
 	process_mapping_set(local_m_vector[i]->first, local_m_vector[i]->second,
 			    *chunk.out_str, *chunk.err_str);
       }
